pyproject.toml
[tool.poetry]
name = "python-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
flask = "^1.1.2"
replit = "^2.0.0"
requests-oauthlib = "^1.3.0"
oauth = "^1.0.1"
"discord.py" = "^1.5.1"
volkswagencarnet = "^4.4.22"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
